Add working code
When splitting a single large file with multiple chapters into multiple files,
the chapters need to be renamed so that their titles make sense for that new
file.

Do this with the default chapter file format in XML.  Deserialize this, create
chapter subsets based on identified episode boundaries, and then fix up the
chapter name and language.  Reserialize this for each episode.

Need to keep the timestamps with respect to the original file because mkvmerge
will adjust these automatically.  Be careful to not modify these at all.