Rewrite process execution utilities
Avoid deadlocks by reading the standard output/error first and then waiting for
the process to exit.

Throw an exception when processes exits with a non-zero exit code.  Make this a
typed exception to enable better error handling.

Return a tuple after running a process that returns the standard error and
standard output as named values.